Restructure access-control package
