+++ .changeset/0c37f0c8/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/access-control", "type": "patch" }], "dependents": [] }
