+++ .changeset/0c37f0c8/changes.md
@@ -1,0 +1,1 @@
+- Restructure internal code
