+++ packages/access-control/tests/access-control.test.js
@@ -3,7 +3,7 @@ import {
   parseFieldAccess,
   validateListAccessControl,
   validateFieldAccessControl,
-} from './index.js';
+} from '../lib/access-control.js';
 describe('Access control package tests', () => {
   test('parseListAccess', () => {
     const statics = [true, false]; // type StaticAccess = boolean;
